feat: Implement tree-sitter runtime and problem matcher for task management
- Added `tree_sitter_runtime.h` to expose a subset of the tree-sitter C API for semantic tokenization, folding regions, and structural outlines.
- Introduced `problem_matcher.h` and `problem_matcher.cpp` to parse compiler output into diagnostics, supporting multiple formats including GCC, MSBuild, and TypeScript.
- Created `task_config.h` and `task_config.cpp` to define and parse task configurations from JSON, mirroring VS Code's task schema.
- Developed `task_runner.h` and `task_runner.cpp` to manage task execution plans, handling dependencies and execution order.
- Added default task definitions for common build and test operations.

+++ tools/ui/common/src/code_assist.cpp
@@ -237,9 +237,9 @@ void SignatureHelpWidget::ShowSignatures(
   }
   signatures_ = signatures;
   active_signature_ =
-      std::clamp(active_signature, 0, signatures_.size() - 1);
+      std::clamp(active_signature, 0, static_cast<int>(signatures_.size()) - 1);
   const int param_count =
-      signatures_[active_signature_].parameter_labels.size();
+      static_cast<int>(signatures_[active_signature_].parameter_labels.size());
   active_parameter_ =
       param_count == 0 ? 0 : std::clamp(active_parameter, 0, param_count - 1);
   Render();
